Move operator tests to own subdirectory
