+++ tests/BUILD.bazel
@@ -125,7 +125,6 @@ go_test(
         "mdev_configuration_allocation_test.go",
         "migration_test.go",
         "nonroot_test.go",
-        "operator_test.go",
         "pausing_test.go",
         "pool_test.go",
         "portforward_test.go",
@@ -197,9 +196,7 @@ go_test(
         "//pkg/virt-handler/node-labeller/util:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//pkg/virt-launcher/virtwrap/converter:go_default_library",
-        "//pkg/virt-operator/resource/apply:go_default_library",
         "//pkg/virt-operator/resource/generate/components:go_default_library",
-        "//pkg/virt-operator/util:go_default_library",
         "//pkg/virtctl/pause:go_default_library",
         "//pkg/virtctl/softreboot:go_default_library",
         "//pkg/virtctl/vm:go_default_library",
@@ -239,6 +236,7 @@ go_test(
         "//tests/monitoring:go_default_library",
         "//tests/network:go_default_library",
         "//tests/numa:go_default_library",
+        "//tests/operator:go_default_library",
         "//tests/performance:go_default_library",
         "//tests/realtime:go_default_library",
         "//tests/reporter:go_default_library",
@@ -249,9 +247,6 @@ go_test(
         "//tests/virtctl:go_default_library",
         "//tests/watcher:go_default_library",
         "//tools/vms-generator/utils:go_default_library",
-        "//vendor/github.com/Masterminds/semver:go_default_library",
-        "//vendor/github.com/evanphx/json-patch:go_default_library",
-        "//vendor/github.com/google/go-github/v32/github:go_default_library",
         "//vendor/github.com/google/goexpect:go_default_library",
         "//vendor/github.com/gorilla/websocket:go_default_library",
         "//vendor/github.com/mitchellh/go-vnc:go_default_library",
@@ -272,7 +267,6 @@ go_test(
         "//vendor/k8s.io/api/policy/v1beta1:go_default_library",
         "//vendor/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1:go_default_library",
         "//vendor/k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
@@ -280,8 +274,6 @@ go_test(
         "//vendor/k8s.io/apimachinery/pkg/fields:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/labels:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/selection:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/intstr:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/json:go_default_library",
@@ -300,7 +292,6 @@ go_test(
         "//vendor/k8s.io/utils/net:go_default_library",
         "//vendor/k8s.io/utils/pointer:go_default_library",
         "//vendor/kubevirt.io/containerized-data-importer-api/pkg/apis/core/v1beta1:go_default_library",
-        "//vendor/kubevirt.io/controller-lifecycle-operator-sdk/api:go_default_library",
         "//vendor/kubevirt.io/qe-tools/pkg/ginkgo-reporters:go_default_library",
     ],
 )
